# Remove unnessary volatile qualifier in Oj.load