When we use volatile qualifier, seems it has some performance penalty on GCC.
The volatile qualifier is unnecessary since the VALUE type object has been kept on the local stack.
(If we manage VALUE on the heap, we need to be careful.)

−       | before  | after   | result
--       | --      | --      | --
GCC      | 4.969M  | 5.165M  | 1.039x
Clang    | 5.054M  | 5.090M  | -


### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 5.18.12-3-MANJARO
  - AMD Ryzen 7 5700G
  - gcc version 12.1.0
  - clang version 14.0.6
  - Ruby 3.1.2

### GCC
#### Before
```
Warming up --------------------------------------
             Oj.load   497.783k i/100ms
Calculating -------------------------------------
             Oj.load      4.969M (± 0.1%) i/s -     49.778M in  10.017042s
```

#### After
```
Warming up --------------------------------------
             Oj.load   518.127k i/100ms
Calculating -------------------------------------
             Oj.load      5.165M (± 0.1%) i/s -     51.813M in  10.031075s
```

### Clang
#### Before
```
Warming up --------------------------------------
             Oj.load   503.712k i/100ms
Calculating -------------------------------------
             Oj.load      5.054M (± 0.1%) i/s -     50.875M in  10.067027s
```

#### After
```
Warming up --------------------------------------
             Oj.load   510.215k i/100ms
Calculating -------------------------------------
             Oj.load      5.090M (± 0.1%) i/s -     51.022M in  10.022973s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json = Oj.dump(42)

Benchmark.ips do |x|
  x.warmup = 5
  x.time = 10

  x.report('Oj.load') do |times|
    i = 0
    while i < times
      Oj.load(json)
      i += 1
    end
  end
end
```