@ohler55 2022-07-29
I'd have to see more tests with large files that cause a GC while parsing. My understanding is the volatile keyword cause the variable not to be put in a register since Ruby does not"see" register variables. Making the VALUE volatile tests the compiler it may change to put it on the stack instead of in a register.