@ohler55 2022-07-30
I'll have to remember `GC.stress`. Looks very useful when testing.